Externalized the config for Pax Exam in the test-suite inside of jqm-tst. Also added this config to the other test files inside of it
